addressed comments: 1. remove metaclient from constructor 2. make INSTANT_STATE_TIMELINE_SERVER_BASED defaults to false 3. cleanup code
